pyproject.toml
[tool.poetry]
name = "enc_phraseological_bot"
version = "0.1.0"
description = ""
authors = ["alexeygalt <alexeygalt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.13.0"
django-environ = "^0.11.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
